Slack APIの整理
